Split SingleStageConverter and PipelineConverter into separate files
Consistent with the existing pattern (Segmentation/MaxMatchSegmentation,
Dict/MarisaDict), each concrete converter subclass moves to its own
.hpp/.cpp pair. Converter.hpp now contains only the abstract Converter
base and ConverterStream. CMakeLists.txt, BUILD.bazel, and
node/node_opencc.gypi are updated accordingly.

Note about OPENCC_EXPORT on SingleStageConverter and PipelineConverter:

Without dllexport on Windows DLL builds, the symbols are absent from the
import library and any code outside the DLL that directly constructs
these classes (including ConversionChainTest) fails to link. OPENCC_EXPORT
is empty on Linux and macOS, so there is no semantic change there.
