+++ node/node_opencc.gypi
@@ -16,12 +16,14 @@
         "../src/Lexicon.cpp",
         "../src/MarisaDict.cpp",
         "../src/MaxMatchSegmentation.cpp",
+        "../src/PipelineConverter.cpp",
         "../src/PluginSegmentation.cpp",
         "../src/PrefixMatch.cpp",
         "../src/ResourceProvider.cpp",
         "../src/Segmentation.cpp",
         "../src/SerializableDict.cpp",
         "../src/SerializedValues.cpp",
+        "../src/SingleStageConverter.cpp",
         "../src/TextDict.cpp",
         "../src/UTF8Util.cpp",
         "../deps/marisa-0.3.1/lib/marisa/agent.cc",
