+++ scripts/build-node-windows-zig.sh
@@ -38,7 +38,9 @@ sources=(
   src/Lexicon.cpp
   src/MarisaDict.cpp
   src/MaxMatchSegmentation.cpp
+  src/PipelineConverter.cpp
   src/PluginSegmentation.cpp
+  src/SingleStageConverter.cpp
   src/PrefixMatch.cpp
   src/ResourceProvider.cpp
   src/Segmentation.cpp
