+++ src/PipelineConverter.cpp
@@ -1,0 +1,41 @@
+/*
+ * Open Chinese Convert
+ *
+ * Copyright 2010-2026 Carbo Kuo and contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "PipelineConverter.hpp"
+
+using namespace opencc;
+
+std::string PipelineConverter::Convert(std::string_view text) const {
+  std::string result(text);
+  for (const ConverterPtr& stage : stages) {
+    result = stage->Convert(result);
+  }
+  return result;
+}
+
+ConversionInspectionResult
+PipelineConverter::Inspect(const std::string& text) const {
+  ConversionInspectionResult result;
+  result.input = text;
+  result.output = Convert(text);
+  return result;
+}
+
+SegmentationPtr PipelineConverter::GetSegmentation() const {
+  return stages.empty() ? nullptr : stages.back()->GetSegmentation();
+}
